Fix bug in TypedSymbolic that does not add constraints correctly when returning from symbolic blocks.
The reasoning in the previous checkin is wrong: the problem is actually the handling of variables that are unannotated or are never assigned from/to annotated variables (i.e., unconstrained variables). For unconstrained variables, returning from the symbolic executor should not add any additional constraints in the case of the $null/$nonnull or any other discrete partial order. Roughly speaking, it's as if the solution for unconstrained variables is top, which is implicitly added to the discrete partial order. In general, returning from the symbolic executor should add the join over the solutions of all paths.

The tests in TypedTopIntegration has also been expanded to cover many combinations of unannotated/non-null annotated and global variables/output arguments/fields. 5 tests are currently failing due to other bugs.