Fix the bug in the unification procedure in UnionQualType.
The problem is due to incorrectly sharing qualifier variables when unifying two types. When unifying a -> b and c -> b, where qualifier variables in a and c are shared with b, some of those shared qualifier variables end up not being unified. The fix is to not share qualifier variables, but to create new ones for b and unify those with a and c.

Also refactored the code a little for organization, to eliminate duplicate calls, and to add some comments.