Refactored TypedSymbolic, moving all conversion functions into SwitchingUtil and making them monadic which makes the code somewhat less messy.
Also fixed the conversion between typed and symbolic, yet again; the previous fix in r8535 was wrong. I just realized that $null annotations are sources of null, and $nonnull annotations are sinks that cannot accept NULL, i.e., $null solutions mark and are converted into may-null pointers, and $nonnull solutions mark must-not-null pointers and are converted into non-null pointers. In the reverse direction, null pointers are converted to $null constraints, but non-null pointers are not converted into any constraints.