Implemented lazy memory initialization.
Types.state.block_to_bytes now maps blocks to deferred bytes, which may be Immediate of bytes, or Deferred of (state -> state * bytes).

MemOp.state__force forces the deferred bytes to be initialized and may update the state in the process. Most of edits are to account for updates to the state.

MemOp.state__add_deferred_block adds a deferred bytes to the memory.

Also took the opportunity to do some style cleanup and refactoring, especially in builtin_function.ml.