Added two features useful for Mix:
Added extra_locals in the symbolic executor state to stash away extra local variables to account for recursion that can occur in the call stack, but is not in the recorded call stack because the execution began "in the middle" of the program.

Added bytes__maybytes_from_list to create a MayBytes tree from a list of bytes.