Added an implementation of "starting in the middle" for Mix by initializing the memory to be consistent with any possible execution beginning at a given point in the program, based on Cil's pointer analysis.
There are still many limitations in the implementation, e.g., lack of support for structures, arrays, malloc. Also, the implementation should technically be usable for applications other than Mix and could potentially be refactored to be part of Otter. For more details, please refer to TODOs and comments in the code, mainly in SwitchingUtil.ml.