Make the varinfo maps conditional and lazy as well. They now map to lval_block deferred rather than to memory_block, which means they can point to several blocks and can be initialized at use.
lval_block can be thought of as a subset of bytes, specifically those that can be resolved to blocks or locations in memory. This actually makes Otter's representation of variables and locations a little closer to the typical lambda calculus plus references, in the way that allows variables to alias. Although basic C variables cannot actually alias, this makes it convenient to set up unknown aliasing relationships prior to "starting in the middle".

I took the opportunity to ditch the memory_frame record with one field varinfo_to_block too, since, well, it has only one field.

Mix is currently broken by this change, and will be fixed when I put in the new alias handling scheme.