Merged MayBytes with IfThenElse by introducing a guard type for the conditional operand of IfThenElse.
Several merging tests currently fail because the condition in IfThenElse is not yet interpreted, which is too conservative. Previously, it was interpreted incompletely in some cases and not at all in others; I've removed the incomplete cases for now.

Also, I put in a big TODO in stp.ml with some refactoring and optimization suggestions.