Add a whole bunch of tests that exercise: - aliasing from typed block  entering typed blocks (symbolic memory initialization); - aliasing from symbolic blocks entering typed blocks; - fixpoint computation; - transitive switching from typed to symbolic to typed.
About the last item: I forgot to mention in the last checkin that I removed the feature that updates qualified type constraints only for symbolic variables that were accessed because it does the wrong thing when entering a fresh typed block. The transitive switching tests exposes this issue.

A total of 14 tests are currently broken since I haven't checked in all features yet.