Extract Otter from Cil into it's own directory otter.
Please "make clean all" from the top-level to clear out stale object files which will confuse the new build system.

There is a README file in otter that provides an overview of how to build, run and develop Otter in the new system. I haven't yet documented the actual Makefile/ocamlbuild build system in detail yet; I'll do so when I get the chance. Please let me know if you really need to know the details.

There's a guaranteedCoverage program which was in cil/src/se/postprocessing that was broken before, and I'm not sure what to do with it. I've put it in otter/util/postprocessing for now.