initialize pointer arguments to be IfThenElse(_,0,malloc()) when SE in the middle
Added 'a conditional constructor ConditionalException. Before that, a job that sees an IfThenElse(b,x,y) will crash if either evaluating x or y crashes. This is because we never fork in eval.ml. If we keep it this way, a solution would be (say y crashes) evaulating ITE(b,x,y) to x assuming b is true. ConditionalException(e) serves as a placeholder for an erroneous leaf.