Add caching for symbolic blocks.
No new tests to add, however, this is also buggy and breaks 4 more tests due to bad interaction with recursion handling for typed blocks, for the same reasons as explained in r9296, and I'll defer the fix as well.

This check-in marks the revision of Mix that was used for the case studies in the PLDI paper.

(Caveat: CilQual/Mix is currently compatible with Otter up to revision r9049 only.)