Automatically (re-)generate API documentation for Otter using ocamldoc.
This subsumes almost all the functionality of otter/doc/Makefile. The API documentation will now be accessible from doc/api/Otter.docdir/index.html, and the module dependency GraphViz file at doc/api/Otter.docdir/Otter.dot. The only difference is that Otter.dot will not be compiled to a PDF using GraphViz dot; instead, I'd suggest using the Graphviz dotty viewer.

This is just a framework to start with. The state of Otter's ocamldoc comments is currently quite miserable, we need to beef it up. Take a look conditional__* in bytes.ml for examples of how to write ocamldoc comments.