Take job_queue out of function call handlers in Interceptors.intercept_function_by_name_internal, and wrap all built-in functions in Builtin_function at declaration rather than at use. Also changed the parameter order by putting job as the first parameter, to be consistent.
job_queue needs to be taken out to be able to wrap built-in functions at declaration, otherwise the compiler will complain about "type variables that cannot be generalized". This is because job_queue is completely unused in Builtin_function, and cannot be given a polymorphic type by the compiler without resorting to eta-expansion. No handlers that correspond to C constructs should ever be manipulating the job_queue anyway.