Remove To_string, refactoring its functionality into a new module TypesPrinter.
The primary change is that TypesPrinter defines only Format-style printers; there are no functions that return strings unlike in To_string. In the few cases where a string is actually needed, FormatPlus.as_string or FormatPlus.sprintf are used instead.

The advantage is that Format-style printers are lazily called by Format.fprintf and family, so strings won't be created just to be discarded if printing is disabled, and the confusing hacks To_string.force_print and To_string.donotprint are no longer needed.