Partial fix for infinite loop issue in Statement.step arising from incorrectly forking states when ConditionalFailureException occurs due to SymbolicPointers.
If a ConditionalFailureException occurs, it should also capture the most recent state from which the failing condition is derived, that can be forked by adding the failing condition to the path condition. Because additional pointers may have been (lazily) initialized, re-running the instruction with the original state will result in a new state that may be inconsistent with the failing condition, and so the failing condition can't be used to fork from the original state.