Separate OtterCore's errors from BackOtter's errors into OtterCore.Errors and BackOtter.BackOtterErrors respectively, and parameterize BasicReporter and OtterPragmaTests on these modules. In particular, this makes BasicReporter and OtterPragmaTests extensible to different front-ends.
The two Errors modules are temporary; eventually, an Otter front-end such as Otter, BackOtter or MultiOtter should define a programmatic interface that combines the interface of Errors and Driver (i.e., the interface provides a way to begin symbolic execution and interpret the results), that is separate from the console interface. Both BackOtter and MultiOtter needs a little more refactoring to make this possible.