Use Rmtmps to remove unreachable functions.
There were two bugs in the implementation of --reachable-functions before. One is minor (I negated the filtering predicate in line 64). But the major bug that can't be fixed easily is that CilCallgraph ignores functions that have no definitions (therefore no fundecs). In this case, Rmtmps should be the way to go.