Fix two bugs in SymbolicPointers test suite, one of which was hiding the other.
The nop(x) idiom to force lazy memory initialization was inadvertently causing additional aliasing (due to context insensitivity), which hid a typo that initialized the wrong pointer. This is replaced by nop(sizeof(x)); sizeof(x) being one of the few expressions that evaluates to bottom in CIL's pointer analysis.