Refactor State.t into an object type such that 'a State.t is supertype of 'a Job.t.
Because 'a Job.t is now a subtype of 'a State.t, they may be used transparently in place of 'a State.t. Many modules, including MemOp and Expression can now directly take a job (instead of having to extract, modify and update states in jobs). If you'd like to enforce particular types, you can write (x : _ #State.t).

Note that subclasses of State.t are not necessarily subtypes due to lazy memory: see the comment documentation for State.t for additional details.