Fix the "Bug: typeOffset: Field on a non-compound" bug in r11132.
The bug is caused by associating the wrong lhost with malloc'ed structs in FunctionJob. Previously, if a pointer points only to a malloc'ed location, the malloc'ed location is initialized using an expression that dereferences the pointer. However, if the pointer points to a field of a malloc'ed struct, dereferencing the pointer does not give the malloc'ed location (it gives the field), and furthermore, it associates the wrong type to the malloc'ed struct (it associates the field's type). The problem occurs when the malloc'ed struct itself contains pointer fields that need to be initialized: the pointer fields are incorrectly referenced via an invalid offset from the wrong type.

The fix associates a list of lhosts with each malloc'ed location instead of dereferencing pointers.