Add loop-elimination that works to speed up my solver variant.
It turns out that there's a serious bug in OLF: in several places, it uses a functional Set to hold mutable data *and* it modified the data, which violates the Set ordering invariant. It's one of a few reasons why collect_ptset_fast doesn't work. My variant doesn't fix this issue but uses a crude workaround.

For coreutils-6.11 chmod in experiments: 14G allocations in 100s without loop-elimination, 5G in 40s with loop-elimination.