Extracted points of interests functions from __otter_main_driver.c into __otter_poi.c, and converted the __FAILURE() call in paste.c to a POI function.
The rationale of these POI functions is that we do not want to remerge coreutils every time we regard some POIs as POIs in a test. Therefore, we create POI functions, and make these functions have either the actual work or no-op based on macros we define when a test is run.