Add the option --backotter-target <file,line[,reason[,reason-argument]>, which provides a richer way to specify targets for BackOtter.
A brief usage guide (since it's currently not very well-documented): file and line are where an Abandoned is expected to occur; reason and reason-argument are arguments to the matchers in Errors and BackOtterErrors, which are also the same arguments to expect_abandoned in PragmaOtterTest. Note that unlike --line-targets, an error has to occur at the target line.