Reverted the changes in r10853. Looks like the path_condition can change upon lazy initialization. For example, with condition like
if(c == 0 && s->next == 0)

where c is symbolic and s is a symbolic pointer with lazy initialization, a condition

If NOT ==(<IfThenElse of s>, 0) Then Bytes(1) Else Bytes(0)

is added to the path_condition after the lazy initialization is forced. Apparently, some special handling of logical expressions is done somewhere in Otter where such condition is added.