Added two fields, instr_id and caller_list, to the class. They have the following properties:
1. instr_id is updated whenever a job's instrList/stmt is updated. Hence it can be thought as one instr_id per execution of instr/stmt at runtime.
2. For jobs a and b, a is the caller of b if a contains a function invocation and immediately leads to the creation of b (that is, b contains an instr/stmt within the function being invoked). caller_list is therefore the list of instr_id's of callers of a job. This list subsumes State.t.callstack.

They are used to keep track of how much time executing an instr/stmt takes. For non-function call instr/stmt, it's just the time running the job. For function call instr, it's the time of summing up all of its callee jobs.

To update the caller_list, whenever the job is updated with a function call instr, the newly generated instr_id is added into the caller_list, and whenever a return stmt is seen, it pops the caller_list. To account for built-in functions that do not have return stmts, force_return is used to unconditionally pop the caller_list.