Prune also the lvals in MemOp.state__deref/MemOp.state__assign, to avoid accessing obsolete lvals.
Currently, there is no visible effect from accessing obsolete lvals, in particular, lvals that fail bounds checking, since the value at the lval is implicitly resized due to the way ImmutableArray works. This check-in is in anticipation for an upcoming check-in which will change ImmutableArray to not resize implicitly and enforce bounds checking.

This leads to a slight 0.5% slowdown on average for do_benchmark.