Fix the BuiltinFunctions.libc_memmove to avoid zero-length reads/writes.
Currently, there is no visible effect from zero-length reads/writes, since ImmutableArray allows the creation of zero-length arrays. This check-in is in anticipation for an upcoming check-in which will change ImmutableArray to not support zero-length arrays, since such arrays are most likely performance or correctness bugs in Otter.

No difference in performance in do_benchmark.